+++ src/Core/Contracts/AddedItemIdResultInterface.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Bitrix24\SDK\Core\Contracts;
+
+interface AddedItemIdResultInterface
+{
+    /**
+     * added entity id
+     *
+     * @return int
+     */
+    public function getId(): int;
+}