+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * add php8 version support
 * change in scope «CRM» Product service and integration tests
+* add `AddedItemIdResultInterface` for batch-queries result
 
 ## 2.0-alpha.2 (31.01.2021)